+++ dataforge-workspace/src/commonMain/kotlin/space/kscience/dataforge/workspace/TaskResult.kt
@@ -1,6 +1,7 @@
 package space.kscience.dataforge.workspace
 
 import space.kscience.dataforge.data.DataSet
+import space.kscience.dataforge.data.forEach
 import space.kscience.dataforge.meta.Meta
 import space.kscience.dataforge.names.Name
 
@@ -23,7 +24,7 @@ public interface TaskResult<out T : Any> : DataSet<T> {
      */
     public val taskMeta: Meta
 
-    override fun traverse(): Sequence<TaskData<T>>
+    override fun iterator(): Iterator<TaskData<T>>
     override fun get(name: Name): TaskData<T>?
 }
 
@@ -34,8 +35,10 @@ private class TaskResultImpl<out T : Any>(
     override val taskMeta: Meta,
 ) : TaskResult<T>, DataSet<T> by dataSet {
 
-    override fun traverse(): Sequence<TaskData<T>> = dataSet.traverse().map {
-        workspace.wrapData(it, it.name, taskName, taskMeta)
+    override fun iterator(): Iterator<TaskData<T>> = iterator {
+        dataSet.forEach {
+            yield(workspace.wrapData(it, it.name, taskName, taskMeta))
+        }
     }
 
     override fun get(name: Name): TaskData<T>? = dataSet.get(name)?.let {