+++ CHANGELOG.md
@@ -16,6 +16,7 @@
 - PartialEnvelope uses `Int` instead `UInt`.
 - `ActiveDataSet` renamed to `DataSource`
 - `selectOne`->`getByType`
+- Data traversal in `DataSet` is done via iterator
 
 ### Deprecated